requirements.txt
python-dotenv==0.17.1
requests==2.25.1
beautifulsoup4==4.9.3
